+++ src/main/java/com/micutne/odik/repository/UserRepository.java
@@ -9,13 +9,13 @@ import org.springframework.stereotype.Repository;
 import java.util.Optional;
 
 @Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    default User findByIdxOrThrow(Long idx) {
+public interface UserRepository extends JpaRepository<User, Integer> {
+    default User findByIdxOrThrow(int idx) {
         return findByIdx(idx)
                 .orElseThrow(() -> new EntityNotFoundException(ErrorCode.USER_NOT_FOUND));
     }
 
-    Optional<User> findByIdx(Long idx);
+    Optional<User> findByIdx(int idx);
 
     default User findByTokenOrThrow(String token) {
         return findByToken(token)
